requirements.txt
Flask==3.1.0
pandas==2.2.3
Pillow==9.0.1
Pillow==11.1.0
Werkzeug==3.1.3
